0.1.0 worked web-sockets input and database connection

+++ src/main/java/web/backend/repositories/DotsRepository.java
@@ -18,7 +18,7 @@ public interface DotsRepository extends PagingAndSortingRepository<DotEntity, Lo
 
     @Query(value = "INSERT INTO dots_table (x, y, r, date, time, owner, creator, area) VALUES (:x, :y, :r, :date, :time, :owner, :creator, :area)")
     @Modifying
-    void addDot(Double x, Double y, Double r, Integer date, Integer time, String owner, String creator, Boolean area);
+    void addDot(Double x, Double y, Double r, Long date, Long time, String owner, String creator, Boolean area);
 
     @Query("SELECT * FROM dots_table WHERE owner = :owner")
     List<DotEntity> getDotsByOwner(String owner);
